+++ .gitignore
@@ -6,3 +6,5 @@
 /html/
 Gemfile.lock
 vendor/
+*.swp
+*.swo
